investigating nan errors in input

+++ snn/data_preprocessing/load_data.py
@@ -41,7 +41,7 @@ def get_event_slice(times, addrs, start_time, end_time, T, size=[128, 128], dt=1
 def chunk_evs_pol(times, addrs, deltat=1000, size=[2, 304, 240], x_max=1, polarity=True):
     t_start = times[0]
     ts = range(t_start, times[-1], deltat)
-    chunks = np.zeros([len(ts)]+size, dtype='int8')
+    chunks = np.zeros([len(ts)]+size)
     idx_start = 0
     idx_end = 0
